refactor: adjust director structure of AddMemberModal, AddonResource, AddonSettings and NoAuthTip
