+++ shell/app/common/__tests__/components/addon-resource/index.test.tsx
