+++ shell/app/common/__tests__/components/no-auth-tip/index.test.tsx
