+++ shell/app/common/components/addon-resource/addon-resource.scss
