+++ shell/app/common/components/addon-settings/addon-settings.scss
