+++ shell/app/modules/msp/env-setting/info/index.tsx
@@ -11,7 +11,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-import { PureAddonSettings } from 'common/components/addon-settings';
+import AddonSettings from 'common/components/addon-settings';
 import { useLoading } from 'core/stores/loading';
 import { Copy, SettingsTabs } from 'common';
 import { isZh } from 'i18n';
@@ -19,6 +19,8 @@ import React from 'react';
 import { useEffectOnce } from 'react-use';
 import mspInfoStore from '../../stores/info';
 import routeInfoStore from 'core/stores/route';
+
+const { PureAddonSettings } = AddonSettings;
 
 const MSComponentInfo = () => {
   const infoList = mspInfoStore.useStore((s) => s.infoList);
